+++ backend/src/test/java/de/bund/digitalservice/ris/caselaw/adapter/InternalPortalPublicationServiceTest.java
@@ -1,6 +1,5 @@
 package de.bund.digitalservice.ris.caselaw.adapter;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -13,11 +12,9 @@ import static org.mockito.Mockito.when;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.bund.digitalservice.ris.caselaw.adapter.caselawldml.CaseLawLdml;
 import de.bund.digitalservice.ris.caselaw.adapter.exception.BucketException;
 import de.bund.digitalservice.ris.caselaw.adapter.exception.LdmlTransformationException;
 import de.bund.digitalservice.ris.caselaw.adapter.exception.PublishException;
-import de.bund.digitalservice.ris.caselaw.adapter.transformer.DocumentationUnitToLdmlTransformer;
 import de.bund.digitalservice.ris.caselaw.domain.CoreData;
 import de.bund.digitalservice.ris.caselaw.domain.DocumentationUnit;
 import de.bund.digitalservice.ris.caselaw.domain.DocumentationUnitRepository;
@@ -29,12 +26,9 @@ import de.bund.digitalservice.ris.caselaw.domain.exception.DocumentationUnitNotE
 import de.bund.digitalservice.ris.caselaw.domain.lookuptable.documenttype.DocumentType;
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 import javax.xml.parsers.DocumentBuilderFactory;
 import net.sf.saxon.TransformerFactoryImpl;
-import org.apache.commons.lang3.StringUtils;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -192,200 +186,5 @@ class InternalPortalPublicationServiceTest {
         .isThrownBy(
             () -> internalPortalPublicationService.publishDocumentationUnit(documentationUnitId))
         .withMessageContaining("Could not save LDML to bucket");
-  }
-
-  @Test
-  @DisplayName("Fallback title test")
-  void documentNumberIsFallbackTitleTest() {
-    String expected =
-        """
-      <akn:header>
-         <akn:p>testDocumentNumber</akn:p>
-      </akn:header>
-     """;
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(
-            testDocumentUnit, documentBuilderFactory);
-    Assertions.assertTrue(ldml.isPresent());
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    Assertions.assertTrue(fileContent.isPresent());
-    Assertions.assertTrue(
-        StringUtils.deleteWhitespace(fileContent.get())
-            .contains(StringUtils.deleteWhitespace(expected)));
-  }
-
-  @Test
-  @DisplayName("Dissenting Opinion test")
-  void dissentingOpinionTest() {
-    String expected =
-        """
-           <akn:block name="Abweichende Meinung">
-              <akn:opinion>
-                 <akn:embeddedStructure>
-                    <akn:p>dissenting test</akn:p>
-                 </akn:embeddedStructure>
-              </akn:opinion>
-           </akn:block>
-           """;
-    DocumentationUnit dissentingCaseLaw =
-        testDocumentUnit.toBuilder()
-            .longTexts(
-                testDocumentUnit.longTexts().toBuilder()
-                    .dissentingOpinion("<p>dissenting test</p>")
-                    .build())
-            .build();
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(
-            dissentingCaseLaw, documentBuilderFactory);
-    Assertions.assertTrue(ldml.isPresent());
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    Assertions.assertTrue(fileContent.isPresent());
-    Assertions.assertTrue(
-        StringUtils.deleteWhitespace(fileContent.get())
-            .contains(StringUtils.deleteWhitespace(expected)));
-  }
-
-  @Test
-  @DisplayName("Headnote test")
-  void headnoteTest() {
-    String expected =
-        """
-            <akn:block name="Orientierungssatz">
-               <akn:embeddedStructure>
-                  <akn:p>headnote test</akn:p>
-               </akn:embeddedStructure>
-            </akn:block>
-           """;
-    DocumentationUnit headnoteCaseLaw =
-        testDocumentUnit.toBuilder()
-            .shortTexts(
-                testDocumentUnit.shortTexts().toBuilder().headnote("<p>headnote test</p>").build())
-            .build();
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(headnoteCaseLaw, documentBuilderFactory);
-    Assertions.assertTrue(ldml.isPresent());
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    Assertions.assertTrue(fileContent.isPresent());
-    Assertions.assertTrue(
-        StringUtils.deleteWhitespace(fileContent.get())
-            .contains(StringUtils.deleteWhitespace(expected)));
-  }
-
-  @Test
-  @DisplayName("OtherHeadnote test")
-  void otherHeadnoteTest() {
-    String expected =
-        """
-            <akn:block name="Sonstiger Orientierungssatz">
-               <akn:embeddedStructure>
-                  <akn:p>other headnote test</akn:p>
-               </akn:embeddedStructure>
-            </akn:block>
-           """;
-    DocumentationUnit otherHeadnoteCaseLaw =
-        testDocumentUnit.toBuilder()
-            .shortTexts(
-                testDocumentUnit.shortTexts().toBuilder()
-                    .otherHeadnote("<p>other headnote test</p>")
-                    .build())
-            .build();
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(
-            otherHeadnoteCaseLaw, documentBuilderFactory);
-    Assertions.assertTrue(ldml.isPresent());
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    Assertions.assertTrue(fileContent.isPresent());
-    Assertions.assertTrue(
-        StringUtils.deleteWhitespace(fileContent.get())
-            .contains(StringUtils.deleteWhitespace(expected)));
-  }
-
-  @Test
-  @DisplayName("Grounds test")
-  void groundTest() {
-    String expected =
-        """
-            <akn:block name="Gründe">
-               <akn:embeddedStructure>
-                  <akn:p>grounds test</akn:p>
-               </akn:embeddedStructure>
-            </akn:block>
-           """;
-    DocumentationUnit groundsCaseLaw =
-        testDocumentUnit.toBuilder()
-            .longTexts(
-                testDocumentUnit.longTexts().toBuilder().reasons("<p>grounds test</p>").build())
-            .build();
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(groundsCaseLaw, documentBuilderFactory);
-    Assertions.assertTrue(ldml.isPresent());
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    Assertions.assertTrue(fileContent.isPresent());
-    Assertions.assertTrue(
-        StringUtils.deleteWhitespace(fileContent.get())
-            .contains(StringUtils.deleteWhitespace(expected)));
-  }
-
-  @Test
-  @DisplayName("OtherLongText without main decision test")
-  void otherLongTextWithoutMainDecisionTest() {
-    String expected =
-        """
-         <akn:decision>
-            <akn:block name="Sonstiger Langtext">
-               <akn:embeddedStructure>
-                  <akn:p>Other long text test</akn:p>
-               </akn:embeddedStructure>
-            </akn:block>
-         </akn:decision>
-         """;
-    DocumentationUnit otherLongTextCaseLaw =
-        testDocumentUnit.toBuilder()
-            .longTexts(
-                testDocumentUnit.longTexts().toBuilder()
-                    .otherLongText("<p>Other long text test</p>")
-                    .build())
-            .build();
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(
-            otherLongTextCaseLaw, documentBuilderFactory);
-    Assertions.assertTrue(ldml.isPresent());
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    Assertions.assertTrue(fileContent.isPresent());
-    Assertions.assertTrue(
-        StringUtils.deleteWhitespace(fileContent.get())
-            .contains(StringUtils.deleteWhitespace(expected)));
-  }
-
-  @Test
-  @DisplayName("Long text with non breaking spaces")
-  void testTransformToLdml_longTextWithNBSP_shouldReplaceItWithUnicode() {
-    String expected =
-        """
-         <akn:decision>
-            <akn:block name="Gründe">
-               <akn:embeddedStructure>
-                  <akn:p>text with non\u00a0breaking\u00a0spaces</akn:p>
-               </akn:embeddedStructure>
-            </akn:block>
-         </akn:decision>
-         """;
-    DocumentationUnit otherLongTextCaseLaw =
-        testDocumentUnit.toBuilder()
-            .longTexts(
-                LongTexts.builder()
-                    .reasons("<p>text with non&nbsp;breaking&nbsp;spaces</p>")
-                    .build())
-            .build();
-
-    Optional<CaseLawLdml> ldml =
-        DocumentationUnitToLdmlTransformer.transformToLdml(
-            otherLongTextCaseLaw, documentBuilderFactory);
-
-    assertThat(ldml).isPresent();
-    Optional<String> fileContent = xmlUtilService.ldmlToString(ldml.get());
-    assertThat(fileContent).isPresent();
-    assertThat(StringUtils.deleteWhitespace(fileContent.get()))
-        .contains(StringUtils.deleteWhitespace(expected));
   }
 }
