+++ backend/src/test/java/de/bund/digitalservice/ris/caselaw/adapter/PublicPortalTransformerTest.java
@@ -1,0 +1,72 @@
+package de.bund.digitalservice.ris.caselaw.adapter;
+
+import de.bund.digitalservice.ris.caselaw.adapter.transformer.PublicPortalTransformer;
+import de.bund.digitalservice.ris.caselaw.domain.CoreData;
+import de.bund.digitalservice.ris.caselaw.domain.DocumentationUnit;
+import de.bund.digitalservice.ris.caselaw.domain.LongTexts;
+import de.bund.digitalservice.ris.caselaw.domain.PreviousDecision;
+import de.bund.digitalservice.ris.caselaw.domain.ShortTexts;
+import de.bund.digitalservice.ris.caselaw.domain.court.Court;
+import de.bund.digitalservice.ris.caselaw.domain.lookuptable.documenttype.DocumentType;
+import java.time.LocalDate;
+import java.util.List;
+import java.util.UUID;
+import javax.xml.parsers.DocumentBuilderFactory;
+import net.sf.saxon.TransformerFactoryImpl;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+class PublicPortalTransformerTest {
+
+  private static DocumentationUnit testDocumentUnit;
+  static XmlUtilService xmlUtilService = new XmlUtilService(new TransformerFactoryImpl());
+
+  private static PublicPortalTransformer subject;
+
+  @BeforeAll
+  static void setUpBeforeClass() {
+    DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();
+
+    subject = new PublicPortalTransformer(documentBuilderFactory);
+
+    PreviousDecision related1 =
+        PreviousDecision.builder()
+            .decisionDate(LocalDate.of(2020, 1, 1))
+            .court(Court.builder().type("Test court type").build())
+            .documentType(DocumentType.builder().label("Test decision type").build())
+            .fileNumber("Test file number")
+            .documentNumber("Test document number 1")
+            .build();
+    PreviousDecision related2 =
+        related1.toBuilder().documentNumber("Test document number 2").build();
+
+    testDocumentUnit =
+        DocumentationUnit.builder()
+            .uuid(UUID.randomUUID())
+            .coreData(
+                CoreData.builder()
+                    .ecli("testecli")
+                    .court(
+                        Court.builder().type("testCourtType").location("testCourtLocation").build())
+                    .documentType(
+                        DocumentType.builder().label("testDocumentTypeAbbreviation").build())
+                    .legalEffect("ja")
+                    .fileNumbers(List.of("testFileNumber"))
+                    .decisionDate(LocalDate.of(2020, 1, 1))
+                    .build())
+            .documentNumber("testDocumentNumber")
+            .longTexts(LongTexts.builder().caseFacts("<p>Example content 1</p>").build())
+            .shortTexts(ShortTexts.builder().build())
+            .previousDecisions(List.of(related1, related2))
+            .build();
+  }
+
+  @Test
+  void transform_withKeywords_shouldNotAddKeywordsToLDML() {
+    // TODO parameterized test for properties which should be excluded from public LDML see
+    // DuplicateCheckIntegrationTest for example
+  }
+}
